Zero out a factor from f.

Zero out a factor from f.

flash_zero_out_factor(data, f, k = 1)

Arguments

data
A flash data object.

f
A flash fit object.

k
Index of factor/loading to zero out.

Details

The factor and loadings of the kth factor of f are made to be zero (except for elements of the factor/loading that are designated to be fixed). This effectively reduces the rank by 1, although the zero factor/loading is kept in f so the number and indexing of factor/loading matrices in f remains the same.