List available available flashcard decks

This function searches for flashcard decks stored in GitHub repositories. By default, the function searches the flashr_decks repo. But other GitHub repos can be used.

To narrow the results, include text in the pattern argument (for example, list_decks(pattern = "r4ds")).

Usage

list_decks(
  pattern = NULL,
  repo = "JeffreyRStevens/flashr_decks",
  quiet = FALSE
)

Arguments

pattern

String pattern to search in list of decks.

repo

GitHub username and repo for deck repository in the format of "username/repository". Default value is "JeffreyRStevens/flashr_decks".

quiet

Logical to prevent list information from printing to console.

Value

Outputs a list of available built-in flashcard decks to the console.

Details

You are welcome to fork the flashr_decks repo and modify or add your own decks. Or you can create your own repo from scratch. Just make sure to place your decks in a directory called decks/ in your root directory. Then set the repo argument to your username and repo (see Examples).

See also

Examples

# View all available decks
list_decks()
#> Available flashcard decks
#> 1. Data types (data_types)
#> 2. Operators (operators)
#> 3. PsyTeachR ADS Ch. 1 (psyteachr_ads1)
#> 4. PsyTeachR ADS Ch. 2 (psyteachr_ads2)
#> 5. PsyTeachR ADS Ch. 3 (psyteachr_ads3)
#> 6. PsyTeachR ADS Ch. 4 (psyteachr_ads4)
#> 7. PsyTeachR ADS Ch. 5 (psyteachr_ads5)
#> 8. PsyTeachR ADS Ch. 7 (psyteachr_ads7)
#> 9. PsyTeachR ADS Ch. 8 (psyteachr_ads8)
#> 10. PsyTeachR ADS Ch. 9 (psyteachr_ads9)
#> 11. PsyTeachR DSRR Ch. 1 (psyteachr_dsrr1)
#> 12. PsyTeachR DSRR Ch. 2 (psyteachr_dsrr2)
#> 13. PsyTeachR DSRR Ch. 3 (psyteachr_dsrr3)
#> 14. PsyTeachR DSRR Ch. 4 (psyteachr_dsrr4)
#> 15. PsyTeachR DSRR Ch. 5 (psyteachr_dsrr5)
#> 16. PsyTeachR DSRR Ch. 6 (psyteachr_dsrr6)
#> 17. PsyTeachR DSRR Ch. 7 (psyteachr_dsrr7)
#> 18. PsyTeachR DSRR Ch. 8 (psyteachr_dsrr8)
#> 19. PsyTeachR DSRR Ch. 9 (psyteachr_dsrr9)
#> 20. PsyTeachR DSRR Ch. 10 (psyteachr_dsrr10)
#> 21. PsyTeachR QFA Ch. 1 (psyteachr_qfa1)
#> 22. PsyTeachR QFA Ch. 2 (psyteachr_qfa2)
#> 23. PsyTeachR QFA Ch. 3 (psyteachr_qfa3)
#> 24. PsyTeachR QFA Ch. 4 (psyteachr_qfa4)
#> 25. PsyTeachR QFA Ch. 5 (psyteachr_qfa5)
#> 26. PsyTeachR QFA Ch. 6 (psyteachr_qfa6)
#> 27. PsyTeachR QFA Ch. 7 (psyteachr_qfa7)
#> 28. PsyTeachR QFA Ch. 8 (psyteachr_qfa8)
#> 29. PsyTeachR QFA Ch. 9 (psyteachr_qfa9)
#> 30. PsyTeachR QFA Ch. 10 (psyteachr_qfa10)
#> 31. PsyTeachR QFA Ch. 11 (psyteachr_qfa11)
#> 32. R4DS Ch. 1 (r4ds1)
#> 33. R4DS Ch. 3 (r4ds3)
#> 34. R4DS Ch. 5 (r4ds5)
#> 35. R4DS Ch. 7 (r4ds7)
#> 36. R4DS Ch. 8 (r4ds8)
#> 37. R4DS Ch. 10 (r4ds10)
#> 38. R4DS Ch. 11 (r4ds11)
#> 39. R4DS Ch. 12 (r4ds12)
#> 40. R4DS Ch. 13 (r4ds13)
#> 41. R4DS Ch. 14 (r4ds14)
#> 42. R4DS Ch. 15 (r4ds15)
#> 43. R4DS Ch. 16 (r4ds16)
#> 44. R4DS Ch. 18 (r4ds18)
#> 45. R4DS Ch. 19 (r4ds19)
#> 46. R4DS Ch. 20 (r4ds20)
#> 47. R4DS Ch. 21 (r4ds21)
#> 48. R4DS Ch. 23 (r4ds23)
#> 49. R4DS Ch. 25 (r4ds25)
#> 50. R4DS Ch. 27 (r4ds27)
#> 51. R4DS Ch. 28 (r4ds28)
#> 52. R4DS Ch. 29 (r4ds29)
#> 53. R4DS Ch. 30 (r4ds30)
#> 54. Vectors (vectors)

# View decks with text matching pattern
list_decks(pattern = "r4ds")
#> Available flashcard decks
#> 1. R4DS Ch. 1 (r4ds1)
#> 2. R4DS Ch. 3 (r4ds3)
#> 3. R4DS Ch. 5 (r4ds5)
#> 4. R4DS Ch. 7 (r4ds7)
#> 5. R4DS Ch. 8 (r4ds8)
#> 6. R4DS Ch. 10 (r4ds10)
#> 7. R4DS Ch. 11 (r4ds11)
#> 8. R4DS Ch. 12 (r4ds12)
#> 9. R4DS Ch. 13 (r4ds13)
#> 10. R4DS Ch. 14 (r4ds14)
#> 11. R4DS Ch. 15 (r4ds15)
#> 12. R4DS Ch. 16 (r4ds16)
#> 13. R4DS Ch. 18 (r4ds18)
#> 14. R4DS Ch. 19 (r4ds19)
#> 15. R4DS Ch. 20 (r4ds20)
#> 16. R4DS Ch. 21 (r4ds21)
#> 17. R4DS Ch. 23 (r4ds23)
#> 18. R4DS Ch. 25 (r4ds25)
#> 19. R4DS Ch. 27 (r4ds27)
#> 20. R4DS Ch. 28 (r4ds28)
#> 21. R4DS Ch. 29 (r4ds29)
#> 22. R4DS Ch. 30 (r4ds30)

# View decks from specific repository
list_decks(repo = "JeffreyRStevens/flashr_decks")
#> Available flashcard decks
#> 1. Data types (data_types)
#> 2. Operators (operators)
#> 3. PsyTeachR ADS Ch. 1 (psyteachr_ads1)
#> 4. PsyTeachR ADS Ch. 2 (psyteachr_ads2)
#> 5. PsyTeachR ADS Ch. 3 (psyteachr_ads3)
#> 6. PsyTeachR ADS Ch. 4 (psyteachr_ads4)
#> 7. PsyTeachR ADS Ch. 5 (psyteachr_ads5)
#> 8. PsyTeachR ADS Ch. 7 (psyteachr_ads7)
#> 9. PsyTeachR ADS Ch. 8 (psyteachr_ads8)
#> 10. PsyTeachR ADS Ch. 9 (psyteachr_ads9)
#> 11. PsyTeachR DSRR Ch. 1 (psyteachr_dsrr1)
#> 12. PsyTeachR DSRR Ch. 2 (psyteachr_dsrr2)
#> 13. PsyTeachR DSRR Ch. 3 (psyteachr_dsrr3)
#> 14. PsyTeachR DSRR Ch. 4 (psyteachr_dsrr4)
#> 15. PsyTeachR DSRR Ch. 5 (psyteachr_dsrr5)
#> 16. PsyTeachR DSRR Ch. 6 (psyteachr_dsrr6)
#> 17. PsyTeachR DSRR Ch. 7 (psyteachr_dsrr7)
#> 18. PsyTeachR DSRR Ch. 8 (psyteachr_dsrr8)
#> 19. PsyTeachR DSRR Ch. 9 (psyteachr_dsrr9)
#> 20. PsyTeachR DSRR Ch. 10 (psyteachr_dsrr10)
#> 21. PsyTeachR QFA Ch. 1 (psyteachr_qfa1)
#> 22. PsyTeachR QFA Ch. 2 (psyteachr_qfa2)
#> 23. PsyTeachR QFA Ch. 3 (psyteachr_qfa3)
#> 24. PsyTeachR QFA Ch. 4 (psyteachr_qfa4)
#> 25. PsyTeachR QFA Ch. 5 (psyteachr_qfa5)
#> 26. PsyTeachR QFA Ch. 6 (psyteachr_qfa6)
#> 27. PsyTeachR QFA Ch. 7 (psyteachr_qfa7)
#> 28. PsyTeachR QFA Ch. 8 (psyteachr_qfa8)
#> 29. PsyTeachR QFA Ch. 9 (psyteachr_qfa9)
#> 30. PsyTeachR QFA Ch. 10 (psyteachr_qfa10)
#> 31. PsyTeachR QFA Ch. 11 (psyteachr_qfa11)
#> 32. R4DS Ch. 1 (r4ds1)
#> 33. R4DS Ch. 3 (r4ds3)
#> 34. R4DS Ch. 5 (r4ds5)
#> 35. R4DS Ch. 7 (r4ds7)
#> 36. R4DS Ch. 8 (r4ds8)
#> 37. R4DS Ch. 10 (r4ds10)
#> 38. R4DS Ch. 11 (r4ds11)
#> 39. R4DS Ch. 12 (r4ds12)
#> 40. R4DS Ch. 13 (r4ds13)
#> 41. R4DS Ch. 14 (r4ds14)
#> 42. R4DS Ch. 15 (r4ds15)
#> 43. R4DS Ch. 16 (r4ds16)
#> 44. R4DS Ch. 18 (r4ds18)
#> 45. R4DS Ch. 19 (r4ds19)
#> 46. R4DS Ch. 20 (r4ds20)
#> 47. R4DS Ch. 21 (r4ds21)
#> 48. R4DS Ch. 23 (r4ds23)
#> 49. R4DS Ch. 25 (r4ds25)
#> 50. R4DS Ch. 27 (r4ds27)
#> 51. R4DS Ch. 28 (r4ds28)
#> 52. R4DS Ch. 29 (r4ds29)
#> 53. R4DS Ch. 30 (r4ds30)
#> 54. Vectors (vectors)